refactor(cli): Phase B args split — run, baseline, bundle, policy
- Extract run.rs (RunArgs, CiArgs)
- Extract baseline.rs (BaselineArgs, BaselineSub, Record, Check, Report)
- Extract bundle.rs (BundleArgs, BundleSub, Create, Verify, ReplayArgs)
- Extract policy.rs (PolicyArgs, PolicyCommand, Validate, Migrate, Fmt)
- Preserve crate::cli::args::* re-exports
